pernosNilton/myDockerAfterLoreto/pdf_table_extractor/requirements.txt
tabula-py==2.9.0
pandas
jpype1
PyPDF2
